rpmsg: char: Migrate to iter versions of read and write
mainline inclusion
from mainline-v6.1-rc6
commit ccf45b18ce89f598c69a0c945ced1635013fc0b1
category: bugfix
issue:#I627JW
CVE:NA

Signed-off-by: gaochao <gaochao49@huawei.com>
---------------------------------------

In order to be able to use the aio interface for writing to a rpmsg_char
the write_iter function must be implemented, so migrate to iter version
for read and write functions.

Regular read and write uses the iter methods if present and is as such
unaffected.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: gaochao <gaochao49@huawei.com>