udf: Get rid of 0-length arrays in struct fileIdentDesc
mainline inclusion
from mainline-v6.1-rc8
commit 979a6e28dd969a2222545001f79566b4bfaf06c0
category: bugfix
issue:#I6287C
CVE:NA

Signed-off-by: gaochao <gaochao49@huawei.com>
---------------------------------------

Get rid of 0-length arrays in struct fileIdentDesc. This requires a bit
of cleaning up as the second variable length array in this structure is
often used and the code abuses the fact that the first two arrays have
the same type and offset in struct fileIdentDesc.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: gaochao <gaochao49@huawei.com>