lib/timerqueue: Rely on rbtree semantics for next timer commit:79edf6d87f86c34e0aed2b8b3566a981f90312ce CVE:CVE-2021-20317 Signed-off-by: wanxiaoqing <wanxiaoqing@huawei.com> ---------------------------
mainline inclusion
from mainline-v5.4-rc1
commit 511885d7061eda3eb1faf3f57dcc936ff75863f1
category: bugfix
bugzilla: NA
CVE: CVE-2021-20317

---------------------------

Simplify the timerqueue code by using cached rbtrees and rely on the tree
leftmost node semantics to get the timer with earliest expiration time.
This is a drop in conversion, and therefore semantics remain untouched.

The runtime overhead of cached rbtrees is be pretty much the same as the
current head->next method, noting that when removing the leftmost node,
a common operation for the timerqueue, the rb_next(leftmost) is O(1) as
well, so the next timer will either be the right node or its parent.
Therefore no extra pointer chasing. Finally, the size of the struct
timerqueue_head remains the same.

Passes several hours of rcutorture.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lkml.kernel.org/r/20190724152323.bojciei3muvfxalm@linux-r8p5
Signed-off-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Reviewed-by: Xie XiuQi <xiexiuqi@huawei.com>
Signed-off-by: wanxiaoqing <wanxiaoqing@huawei.com>