lkdtm/bugs: Check for the NULL pointer after calling kmalloc
mainline inclusion
from mainline-v6.2-rc1
commit 4a9800c81d2f34afb66b4b42e0330ae8298019a2
category: bugfix
issue:#I6718C
CVE: CVE-2022-3104

Signed-off-by: gaochao <gaochao49@huawei.com>
---------------------------------------

As the possible failure of the kmalloc(), the not_checked and checked
could be NULL pointer.
Therefore, it should be better to check it in order to avoid the
dereference of the NULL pointer.
Also, we need to kfree the 'not_checked' and 'checked' to avoid
the memory leak if fails.
And since it is just a test, it may directly return without error
number.

Fixes: ae2e1aad3e48 ("drivers/misc/lkdtm/bugs.c: add arithmetic overflow and array bounds checks")
Signed-off-by: Jiasheng Jiang <jiasheng@iscas.ac.cn>
Acked-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Kees Cook <keescook@chromium.org>
Link: https://lore.kernel.org/r/20220120092936.1874264-1-jiasheng@iscas.ac.cn
Signed-off-by: gaochao <gaochao49@huawei.com>