efi: capsule-loader: Fix use-after-free in efi_capsule_write
stable inclusion
from stable-v4.19.256
commit 021805af5bedeafc76c117fc771c100b358ab419
category: bugfix
issue:#I62LR5
CVE:CVE-2022-40307

Signed-off-by: gaochao <gaochao49@huawei.com>
---------------------------------------

commit 9cb636b5f6a8cc6d1b50809ec8f8d33ae0c84c95 upstream.

A race condition may occur if the user calls close() on another thread
during a write() operation on the device node of the efi capsule.

This is a race condition that occurs between the efi_capsule_write() and
efi_capsule_flush() functions of efi_capsule_fops, which ultimately
results in UAF.

So, the page freeing process is modified to be done in
efi_capsule_release() instead of efi_capsule_flush().

Cc: <stable@vger.kernel.org> # v4.9+
Signed-off-by: Hyunwoo Kim <imv4bel@gmail.com>
Link: https://lore.kernel.org/all/20220907102920.GA88602@ubuntu/
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: gaochao <gaochao49@huawei.com>