proc: avoid integer type confusion in get_proc_long
stable inclusion
from stable-v4.19.268
commit fe84d7f0cb66d150de094fba461f0cb5d5b12c85
category: bugfix
issue:#I6D34S
CVE: CVE-2022-4378

Signed-off-by: wanxiaoqing <wanxiaoqing@huawei.com>
---------------------------------------

commit e6cfaf34be9fcd1a8285a294e18986bfc41a409c upstream.

proc_get_long() is passed a size_t, but then assigns it to an 'int'
variable for the length.  Let's not do that, even if our IO paths are
limited to MAX_RW_COUNT (exactly because of these kinds of type errors).

So do the proper test in the rigth type.

Reported-by: Kyle Zeng <zengyhkyle@gmail.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: wanxiaoqing <wanxiaoqing@huawei.com>